DTO moved to main folder
