+++ siforovProj/src/main/java/api/PrivateCurrencyCourseDTO.java
@@ -1,4 +1,4 @@
-package apiTests;
+package api;
 
 import com.google.gson.annotations.SerializedName;
